requirements.txt
torch==1.11.0
pytorch-lightning==1.6.4
antlr4-python3-runtime==4.9.3
cachetools==4.2.4
carbontracker==1.1.6
geocoder==1.38.1
google-auth==1.35.0
hydra-core==1.2.0
icd10-cm==0.0.5
liac-arff==2.5.0
lightly==1.2.22
lightly-utils==0.0.2
lightning-bolts==0.5.0
matplotlib-venn==0.11.7
minio==7.1.9
omegaconf==2.2.2
opencv-python==4.6.0.66
openml==0.12.2
pyarrow==8.0.0
ratelim==0.1.6
simple-icd-10==2.0.1
tabpfn==0.1.8
torch-tb-profiler==0.4.0
xgboost==1.7.4
xmltodict==0.13.0
albumentations==1.3.1
einops==0.7.0
timm==0.9.12
numpy==1.23.5
streamlit==1.42.0